Fix LargeInt/Integer mismatch: sick_leave_rates.min/max_years -> BIGINT
SQLite's manifest typing lets an INTEGER column hold values wider than
32 bits (e.g. a large "no upper bound" sentinel in the top experience
bracket), so FireDAC's SQLite driver reported these columns as
LargeInt. PostgreSQL's plain INTEGER is a fixed 4-byte type, so the
persistent LargeInt fields from the old SQLite-based dataset design
now mismatch. Switch both columns to BIGINT and add a migration for
databases that already applied the old INTEGER-based schema.

+++ database/postgres/001_initial_schema.sql
@@ -163,8 +163,14 @@ CREATE TABLE IF NOT EXISTS sick_leave_journal (
 
 CREATE TABLE IF NOT EXISTS sick_leave_rates (
     id         INTEGER PRIMARY KEY,
-    min_years  INTEGER,
-    max_years  INTEGER,
+    -- BIGINT, не INTEGER: у верхней ступени стажа обычно стоит
+    -- условная "бесконечность" (например 999999999) как значение
+    -- max_years — SQLite хранит INTEGER произвольной величины и
+    -- FireDAC для такой колонки определял поле как LargeInt; обычный
+    -- 4-байтовый INTEGER в Postgres даёт EDatabaseError "Type mismatch
+    -- ... expecting: LargeInt actual: Integer".
+    min_years  BIGINT,
+    max_years  BIGINT,
     percent    DOUBLE PRECISION
 );
